address some of the comments

+++ projects/rocsparse/library/src/level3/coomm_device_atomic.h
@@ -87,7 +87,7 @@ namespace rocsparse
                         {
                             rocsparse::atomic_add(
                                 dense_C,
-                                static_cast<int>((colB + p * WF_SIZE) * ldc + current_row),
+                                static_cast<int64_t>((colB + p * WF_SIZE) * ldc + current_row),
                                 dense_C_size,
                                 static_cast<C>(alpha * sum[p]));
                         }
@@ -98,7 +98,7 @@ namespace rocsparse
                         {
                             rocsparse::atomic_add(
                                 dense_C,
-                                static_cast<int>(current_row * ldc + colB + p * WF_SIZE),
+                                static_cast<int64_t>(current_row * ldc + colB + p * WF_SIZE),
                                 dense_C_size,
                                 static_cast<C>(alpha * sum[p]));
                         }
@@ -140,7 +140,7 @@ namespace rocsparse
                 {
                     rocsparse::atomic_add(
                         dense_C,
-                        static_cast<int>((colB + p * WF_SIZE) * ldc + current_row),
+                        static_cast<int64_t>((colB + p * WF_SIZE) * ldc + current_row),
                         dense_C_size,
                         static_cast<C>(alpha * sum[p]));
                 }
@@ -149,10 +149,11 @@ namespace rocsparse
             {
                 for(uint32_t p = 0; p < LOOPS; p++)
                 {
-                    rocsparse::atomic_add(dense_C,
-                                          static_cast<int>(current_row * ldc + colB + p * WF_SIZE),
-                                          dense_C_size,
-                                          static_cast<C>(alpha * sum[p]));
+                    rocsparse::atomic_add(
+                        dense_C,
+                        static_cast<int64_t>(current_row * ldc + colB + p * WF_SIZE),
+                        dense_C_size,
+                        static_cast<C>(alpha * sum[p]));
                 }
             }
         }
