+++ projects/rocsparse/library/src/level3/csrmm_device_nnz_split.h
@@ -504,7 +504,7 @@ namespace rocsparse
                         {
                             rocsparse::atomic_add(
                                 dense_C,
-                                static_cast<int>((colB + p * WF_SIZE) * ldc + current_row),
+                                static_cast<int64_t>((colB + p * WF_SIZE) * ldc + current_row),
                                 dense_C_size,
                                 static_cast<C>(alpha * sum[p]));
                         }
@@ -515,7 +515,7 @@ namespace rocsparse
                         {
                             rocsparse::atomic_add(
                                 dense_C,
-                                static_cast<int>(current_row * ldc + colB + p * WF_SIZE),
+                                static_cast<int64_t>(current_row * ldc + colB + p * WF_SIZE),
                                 dense_C_size,
                                 static_cast<C>(alpha * sum[p]));
                         }
@@ -542,7 +542,7 @@ namespace rocsparse
                 {
                     rocsparse::atomic_add(
                         dense_C,
-                        static_cast<int>((colB + p * WF_SIZE) * ldc + current_row),
+                        static_cast<int64_t>((colB + p * WF_SIZE) * ldc + current_row),
                         dense_C_size,
                         static_cast<C>(alpha * sum[p]));
                 }
@@ -551,10 +551,11 @@ namespace rocsparse
             {
                 for(unsigned int p = 0; p < LOOPS; p++)
                 {
-                    rocsparse::atomic_add(dense_C,
-                                          static_cast<int>(current_row * ldc + colB + p * WF_SIZE),
-                                          dense_C_size,
-                                          static_cast<C>(alpha * sum[p]));
+                    rocsparse::atomic_add(
+                        dense_C,
+                        static_cast<int64_t>(current_row * ldc + colB + p * WF_SIZE),
+                        dense_C_size,
+                        static_cast<C>(alpha * sum[p]));
                 }
             }
         }
@@ -635,14 +636,14 @@ namespace rocsparse
                         if(order_C == rocsparse_order_column)
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(colB * ldc + current_row),
+                                                  static_cast<int64_t>(colB * ldc + current_row),
                                                   dense_C_size,
                                                   static_cast<C>(alpha * sum));
                         }
                         else
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(current_row * ldc + colB),
+                                                  static_cast<int64_t>(current_row * ldc + colB),
                                                   dense_C_size,
                                                   static_cast<C>(alpha * sum));
                         }
@@ -693,17 +694,19 @@ namespace rocsparse
                     {
                         if(order_C == rocsparse_order_column)
                         {
-                            rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>((l + swid) * ldc + current_row),
-                                                  dense_C_size,
-                                                  static_cast<C>(alpha * sum));
+                            rocsparse::atomic_add(
+                                dense_C,
+                                static_cast<int64_t>((l + swid) * ldc + current_row),
+                                dense_C_size,
+                                static_cast<C>(alpha * sum));
                         }
                         else
                         {
-                            rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(current_row * ldc + (l + swid)),
-                                                  dense_C_size,
-                                                  static_cast<C>(alpha * sum));
+                            rocsparse::atomic_add(
+                                dense_C,
+                                static_cast<int64_t>(current_row * ldc + (l + swid)),
+                                dense_C_size,
+                                static_cast<C>(alpha * sum));
                         }
                     }
                 }
@@ -717,17 +720,19 @@ namespace rocsparse
                     {
                         if(order_C == rocsparse_order_column)
                         {
-                            rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>((l + swid) * ldc + current_row),
-                                                  dense_C_size,
-                                                  static_cast<C>(alpha * sum));
+                            rocsparse::atomic_add(
+                                dense_C,
+                                static_cast<int64_t>((l + swid) * ldc + current_row),
+                                dense_C_size,
+                                static_cast<C>(alpha * sum));
                         }
                         else
                         {
-                            rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(current_row * ldc + (l + swid)),
-                                                  dense_C_size,
-                                                  static_cast<C>(alpha * sum));
+                            rocsparse::atomic_add(
+                                dense_C,
+                                static_cast<int64_t>(current_row * ldc + (l + swid)),
+                                dense_C_size,
+                                static_cast<C>(alpha * sum));
                         }
                     }
                 }
